add missing package.json files entries
When present, `package.json`'s `files` array acts as a filter, only
including the file-types which are explicitly listed when the package is
installed as a dependency. Without entries for `icons` and
`translations`, those elements were missing from the bundle when the
build was constructed outside the context of a workspace `ui-marccat`
had been cloned into.

See https://docs.npmjs.com/files/package.json#files for details