clean up peer-deps
* increment `@folio/stripes` from `v3` to `v4`
* increment package version from `2.0.0` to `3.0.0` which should have
happened when `react-intl` was bumped to `v4`
* move `react-intl` to peerDeps
* remove `react-router` and `react-router-dom` as direct deps since they
are already correctly specified as peers
* resolve the conflicting versions of `react-router` to the peerDeps
version

The direct-dep on `react-intl` is preventing platform-complete#snapshot
from building at present because of a bug in the 4.7 line. Removing the
direct dep here means we can fix this with a single change in the
platform.